(maint) Add error message and tidy up namespace
Prior to this commit there was no error message when the project was not
passed a file argument and the config and nrepl-server had to be
declared. This commit adds an error message and passes config to the
make-webapp wrapper function and its contents are used by the iternal
functions.