(maint) Add connection pooling and logging
Prior to this commit dujour did not pool connections to the postgresql
database. This commit adds bonecp connection pooling similar to puppetdb
to make the initial import of dujour checkin data to the database much
faster and also adds logging in the form of log4j, again similar to the
way that puppetdb is configured.